% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R, R/qcs_protocol_id.R
\name{metadata}
\alias{metadata}
\alias{qcs_dataset_template}
\alias{qcs_dataset_column_names}
\alias{assert_is_qcs_dataset}
\alias{assert_is_qcs_protocol_id}
\title{JRC-ENCR QCS Metadata}
\usage{
qcs_dataset_template(qcs_protocol_id, assertion_type = NULL)

qcs_dataset_column_names(qcs_protocol_id)

assert_is_qcs_dataset(
  x,
  x_nm = NULL,
  call = NULL,
  assertion_type = NULL,
  qcs_protocol_id
)

assert_is_qcs_protocol_id(x, x_nm = NULL, call = NULL, assertion_type = NULL)
}
\arguments{
\item{qcs_protocol_id}{\verb{[character, integer]} (no default)
\itemize{
\item \code{integer}: Used as-is. You have to study what the \code{.jar} does with that.
\item \code{character}: Take 11, 13, 15, or 17 for \code{"incidence"}, \code{"mortality"},
\code{"population"}, and \code{"lifetable"}, respectively.
}}

\item{assertion_type}{See \verb{[dbc::handle_arg_assertion_type]}}

\item{x}{\verb{[any class]} (no default)

R object to perform assertion on.}

\item{x_nm}{See \verb{[dbc::handle_arg_x_nm]}}

\item{call}{See \verb{[dbc::handle_arg_call]}}
}
\description{
Tools to access metadata such as dataset names and column names of a dataset.
}
\details{
\verb{[ecnrqcs::qcs_dataset_template]} returns a \code{data.table} with one row which
has the required columns in their required format for the given
\code{dataset_name}. Column names and classes taken from JRC-ENCR User
Compendium 2.0.

Replaced argument \code{dataset_name} with \code{qcs_protocol_id}.

\verb{[ecnrqcs::qcs_dataset_column_names]} returns a character string vector
of column names for the given \code{qcs_protocol_id}. This function wraps
\verb{[ecnrqcs::qcs_dataset_template]}.
}
