% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcs.R
\name{qcs}
\alias{qcs}
\title{JRC-ENCR QCS}
\usage{
qcs(
  dataset,
  dataset_name,
  qcs_dir_path,
  dataset_file_path = NULL,
  clean = c("input", "output", "both", "neither")[3L],
  write_arg_list = NULL,
  run_arg_list = NULL,
  read_arg_list = NULL,
  assertion_type = "input"
)
}
\arguments{
\item{dataset}{\verb{[data.frame]} (mandatory)

Column requirements vary depending on \code{dataset_name}.
Refer to the JRC-ENCR QCS User Compendium.}

\item{dataset_name}{\verb{[character]} (no default)

one of the following:
\itemize{
\item "incidence"
\item "mortality"
\item "population"
\item "lifetable"
}}

\item{qcs_dir_path}{\verb{[character]} (no default)

Path to root directory of the JRC-ENCR QCS.}

\item{dataset_file_path}{\verb{[NULL, character]} (default \code{NULL})

\code{dataset} needs to be written to hard drive for use by JRC-ENCR QCS.
You can (optionally) specify the path where to write \code{dataset} via this
arg.
\itemize{
\item \code{NULL}: \code{dataset} will be written into a temporary file given by
\code{tempfile(fileext = ".csv")}.
\item \code{character}: \code{dataset} will be written here.
}}

\item{clean}{\verb{[logical]} (default \code{"both"})

One of the following:
\itemize{
\item \code{"input"}  : Delete input file when this function no longer needs it.
\item \code{"output"} : Delete output files when this function no longer need them.
\item \code{"both"}   : Delete both input and output files.
\item \code{"neither"}: Don't delete any files.
}}

\item{write_arg_list}{\verb{[NULL, list]} (default \code{NULL})

Optional, additional arguments passed to \verb{[encrqcs::qcs_write_dataset]}
if a list.}

\item{run_arg_list}{\verb{[NULL, list]} (default \code{NULL})

Optional, additional arguments passed to \verb{[encrqcs::qcs_run]}
if a list.}

\item{read_arg_list}{\verb{[NULL, list]} (default \code{NULL})

Optional, additional arguments passed to \verb{[encrqcs::qcs_read_results]}
if a list.}

\item{assertion_type}{See e.g. \verb{[dbc::report_to_assertion]}.}
}
\description{
Run JRC-ENCR QCS and read results into R.
}
