% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{qcs_cache}
\alias{qcs_cache}
\alias{qcs_cache_dir_path}
\alias{qcs_cache_metadata_file_path}
\alias{qcs_cache_metadata_read}
\alias{qcs_cache_metadata_write}
\alias{qcs_cache_metadata_update}
\alias{qcs_cache_dataset_file_hash}
\title{\code{encrqcs} Cache}
\usage{
qcs_cache_dir_path(qcs_dir_path)

qcs_cache_metadata_file_path(qcs_dir_path)

qcs_cache_metadata_read(qcs_dir_path)

qcs_cache_metadata_write(cache_metadata_dt, qcs_dir_path)

qcs_cache_metadata_update(
  qcs_dir_path,
  qcs_protocol_id,
  run_id,
  hash = NULL,
  dataset_file_path = NULL
)

qcs_cache_dataset_file_hash(dataset_file_path)
}
\arguments{
\item{qcs_dir_path}{\verb{[character]} (no default)

Path to root directory of the JRC-ENCR QCS.}

\item{cache_metadata_dt}{\verb{[data.table]} (no default)

Table of metadata to write into cache. Must have these columns:
\code{c("metadata_datetime", "hash", "qcs_protocol_id", "run_id")}.}

\item{qcs_protocol_id}{\verb{[character, integer]} (no default)
\itemize{
\item \code{integer}: Used as-is. You have to study what the \code{.jar} does with that.
\item \code{character}: Take 11, 13, 15, or 17 for \code{"incidence"}, \code{"mortality"},
\code{"population"}, and \code{"lifetable"}, respectively.
}}

\item{run_id}{\verb{[integer]} (no default)

One of the values of column \code{RUN_ID} in \code{temp/validation_run.csv} under
the \code{qcs_dir_path}.}

\item{hash}{\verb{[character, NULL]} (default \code{NULL})

Optional hash (checksum) of file at \code{dataset_file_path}. See
\verb{[qcs_cache_dataset_file_hash]}.
\itemize{
\item \code{NULL}: \code{hash} is computed first on \code{dataset_file_path}.
\item \code{character}: Use this hash.
}}

\item{dataset_file_path}{\verb{[NULL, character]} (default \code{NULL})

Optional path to your dataset.
\itemize{
\item \code{character}: Path to an existing file.
This should be the dataset you want to use.
E.g. \code{"C:/my/dir/mydataset.csv"}.
\item \code{NULL}: This argument is not used.
}}
}
\description{
Functions to make use of the cache used by \code{encrqcs}.
}
\section{Functions}{

\strong{encrqcs::qcs_cache_dir_path}

\code{encrqcs::qcs_cache_dir_path} returns the path to the cache dir used by
\code{encrqcs} given \code{qcs_dir_path} where the JRC-ENCR QCS is located.
It is simply \code{sprintf("\%s/temp/encrqcs/", qcs_dir_path)}.


\strong{encrqcs::qcs_cache_metadata_file_path}

\code{encrqcs::qcs_cache_metadata_file_path} returns the path to metadata file
in the cache.
It is simply \code{sprintf("\%s/metadata.csv", qcs_cache_dir_path(qcs_dir_path))}.


\strong{encrqcs::qcs_cache_metadata_read}

\code{encrqcs::qcs_cache_metadata_read} reads (with \verb{[data.table::fread]})
the file at \code{encrqcs::qcs_cache_metadata_file_path(qcs_dir_path)}. If it
does not exist, an empty \code{data.table} is returned. Currently
the metadata file maps \code{hash} to \code{run_id}, where the former is computed on
a dataset and \code{run_id} corresponds to column \code{RUN_ID} in
\code{temp/validation_run.csv}. This maps a dataset to a previous run reliably.


\strong{encrqcs::qcs_cache_metadata_write}

\code{encrqcs::qcs_cache_metadata_write} writes \code{cache_metadata_dt} with
\verb{[data.table::fwrite]} into
\code{encrqcs::qcs_cache_metadata_file_path(qcs_dir_path)}.


\strong{encrqcs::qcs_cache_metadata_update}

\code{encrqcs::qcs_cache_metadata_update} the metadata file with
\code{encrqcs::qcs_cache_metadata_read}, adds a new record into it,
and re-writes the metadata with \code{encrqcs::qcs_cache_metadata_write}.


\strong{encrqcs::qcs_cache_dataset_file_hash}

\code{encrqcs::qcs_cache_dataset_file_hash} computes the hash using
\verb{[digest::digest]} on an on-disk dataset. It uses
\code{algo = "md5"}.
}

\section{News for version 1.0.0}{


New function \code{encrqcs::qcs_cache_dir_path}.



New function \code{encrqcs::qcs_cache_metadata_file_path}.



New function \code{encrqcs::qcs_cache_metadata_read}.



New function \code{encrqcs::qcs_cache_metadata_write}.



New function \code{encrqcs::qcs_cache_metadata_update}.



New function \code{encrqcs::qcs_cache_dataset_file_hash}.
}

