% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{qcs_clean}
\alias{qcs_clean}
\alias{qcs_clean_output}
\alias{qcs_clean_output_all}
\title{Clean Output}
\usage{
qcs_clean_output(qcs_protocol_id, qcs_dir_path)

qcs_clean_output_all(qcs_dir_path)
}
\arguments{
\item{qcs_protocol_id}{\verb{[character, integer]} (no default)
\itemize{
\item \code{integer}: Used as-is. You have to study what the \code{.jar} does with that.
\item \code{character}: Take 11, 13, 15, or 17 for \code{"incidence"}, \code{"mortality"},
\code{"population"}, and \code{"lifetable"}, respectively.
}}

\item{qcs_dir_path}{\verb{[character]} (no default)

Path to root directory of the JRC-ENCR QCS.}
}
\description{
Functions to clean QCS output from mass memory.
}
\section{Functions}{

\strong{encrqcs::qcs_clean_output}

\code{encrqcs::qcs_clean_output} performs the following steps:
\itemize{
\item Read a metadata .csv from the \code{qcs_dir_path}. This contains information
about every previous run of QCS.
\item Detect the latest run in the metadata with the correct \code{qcs_protocol_id}.
\item Delete the entire directory containing summary data, e.g.
\verb{C:/path/to/qcs/output/incidence}.
\item Delete every .csv file of the detected latest run in
\verb{C:/path/to/qcs/temp}.
\item Remove the latest run from the metadata and write it back, except if
if it was the only run. Then delete that file as well.
}


\strong{encrqcs::qcs_clean_output_all}

\code{encrqcs::qcs_clean_output_all} deletes the subdirectories
\code{output} and \code{temp} under \code{qcs_dir_path}.
}

\section{News for version 0.8.0}{


New function \code{encrqcs::qcs_clean_output}.



New function \code{encrqcs::qcs_clean_output_all}.
}

