% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{qcs_column_specifications}
\alias{qcs_column_specifications}
\title{Column Specifications}
\usage{
qcs_column_specifications(dataset_name = "incidence")
}
\arguments{
\item{dataset_name}{\verb{[character]} (mandatory, no default)

one of the following:
\itemize{
\item "incidence"
}}
}
\description{
Retrieve a table of column specifications for the reqested dataset name.
}
