% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{qcs_write_dataset}
\alias{qcs_write_dataset}
\title{JRC-ENCR QCS Input and Output}
\usage{
qcs_write_dataset(
  dataset,
  qcs_protocol_id,
  file_path,
  fwrite_arg_list = NULL,
  assertion_type = NULL
)
}
\arguments{
\item{dataset}{\verb{[data.frame]} (mandatory)

Column requirements vary depending on \code{qcs_protocol_id}.
Refer to the JRC-ENCR QCS User Compendium.}

\item{qcs_protocol_id}{\verb{[character, integer]} (no default)
\itemize{
\item \code{integer}: Used as-is. You have to study what the \code{.jar} does with that.
\item \code{character}: Take 11, 13, 15, or 17 for \code{"incidence"}, \code{"mortality"},
\code{"population"}, and \code{"lifetable"}, respectively.
}}

\item{file_path}{\verb{[character]} (no default)

path to write file to}

\item{fwrite_arg_list}{\verb{[NULL, list]} (default \code{NULL})

Additional arguments passed to \verb{[data.table::fwrite]} if a list. Arguments
\code{x}, \code{file}, and \code{sep} are determined internally and cannot be changed.}

\item{assertion_type}{See e.g. \verb{[dbc::report_to_assertion]}.}
}
\value{
\verb{[encrqcs::qcs_write_dataset]} always returns \code{NULL} invisibly. As a
side effect it writes \code{dataset} to disk.
}
\description{
Read and write .csv files for / created by the JRC-ENCR QCS.
}
\details{
\verb{[encrqcs::qcs_write_dataset]} performs simple checks on its input and
writes \code{dataset} on the disk using \verb{[data.table::fwrite]}.
}
