% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_qcs_file}
\alias{write_qcs_file}
\title{JRC-ENCR QCS Input and Output}
\usage{
write_qcs_file(dataset, dataset_name = "incidence", file_path, ...)
}
\arguments{
\item{dataset}{\verb{[data.frame]} (mandatory, no default)

e.g. incidence (cancer record) dataset}

\item{dataset_name}{\verb{[character]} (mandatory, no default)

one of the following:
\itemize{
\item "incidence"
}}

\item{file_path}{\verb{[character]} (mandatory, no default)

path to write file to}

\item{...}{additional arguments passed to \link[data.table:fwrite]{data.table::fwrite}}
}
\description{
Read and write .csv files for / created by the JRC-ENCR QCS.
}
