% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encrqcs.R
\docType{package}
\name{encrqcs}
\alias{encrqcs-package}
\alias{encrqcs}
\title{\code{encrqcs}: JRC-ENCR QCS Tools}
\description{
encrqcs

\subsection{Installation}{

You can install the release of encrqcs from
\href{https://github.com/FinnishCancerRegistry/encrqcs}{GitHub} with:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# install.packages("devtools")
devtools::install_github("finnishcancerregistry/encrqcs@release")
}\if{html}{\out{</div>}}
}

\subsection{Example}{

Run incidence checks using JRC-ENCR QCS:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{results <- encrqcs::qcs_run(
  my_incidence, "incidence", "path/to/qcs/root/dir/"
)
}\if{html}{\out{</div>}}
}
}
\section{News}{

\subsection{News for version 1.0.0}{
\subsection{encrqcs::qcs_cache_dataset_file_hash}{

New function \code{encrqcs::qcs_cache_dataset_file_hash}.
}

\subsection{encrqcs::qcs_cache_dir_path}{

New function \code{encrqcs::qcs_cache_dir_path}.
}

\subsection{encrqcs::qcs_cache_metadata_file_path}{

New function \code{encrqcs::qcs_cache_metadata_file_path}.
}

\subsection{encrqcs::qcs_cache_metadata_read}{

New function \code{encrqcs::qcs_cache_metadata_read}.
}

\subsection{encrqcs::qcs_cache_metadata_update}{

New function \code{encrqcs::qcs_cache_metadata_update}.
}

\subsection{encrqcs::qcs_cache_metadata_write}{

New function \code{encrqcs::qcs_cache_metadata_write}.
}

\subsection{encrqcs::qcs_read_results}{

\code{encrqcs::qcs_read_results} argument \code{dataset_name} removed.
}

\subsection{encrqcs::qcs_read_results}{

\code{encrqcs::qcs_read_results} gains arguments \code{hash} and \code{dataset_file_path},
both \code{NULL} by default.
}

\subsection{encrqcs::qcs_run}{

\code{encrqcs::qcs_call} gains caching ability. It now saves the hash of every
dataset it has run through QCS and skips the run if the same dataset has
already been run previously.
}

}

\subsection{News for version 0.8.1}{
\subsection{encrqcs}{

New release to trigger github action. No changes in code.
}

}

\subsection{News for version 0.8.0}{
\subsection{encrqcs::qcs_clean_output}{

New function \code{encrqcs::qcs_clean_output}.
}

\subsection{encrqcs::qcs_clean_output_all}{

New function \code{encrqcs::qcs_clean_output_all}.
}

\subsection{encrqcs::qcs_read_results}{

\code{encrqcs::qcs_read_results} revamp: it now reads a different and more usable
set of results than before.
}

\subsection{encrqcs::qcs_run}{

\code{encrqcs::qcs_run} cleaning of output now more thorough.
}

}

\subsection{News for version 0.7.0}{
\subsection{encrqcs::qcs_read_results}{

Fixed \code{encrqcs::qcs_read_results}: it now also reads e.g.
\code{QCS-Incidence-Output-Summary.txt}. The regex used to detect files to read
in the results dir was \code{"output([.]txt)|([.]csv)"} and is now
\code{"output.*([.]txt)|([.]csv)"} (added \verb{.*}).
}

}

\subsection{News for version 0.6.0}{
\subsection{encrqcs::qcs_run}{

\code{encrqcs::qcs_call} gains argument \code{optional_steps}.
}

\subsection{encrqcs::qcs_run}{

\code{encrqcs::qcs_run} now enforces the correct order of columns in \code{dataset}.
}

\subsection{encrqcs::qcs_run}{

Replace output element \code{run_log} with \code{qcs_call_output}, the whole output
object of \code{encrqcs::qcs_call}.
}

\subsection{encrqcs::qcs_write_dataset}{

\code{encrqcs::qcs_write} now enforces the correct order of columns in \code{dataset}.
}

}

\subsection{News for version 0.5.0}{
\subsection{encrqcs::qcs_run}{

Removed override of \code{system2_arg_list} within \code{encrqcs::qcs_run}. So now
the user can pass \code{system2_arg_list} to the call of \code{encrqcs::qcs_call}.
}

}

\subsection{News for version 0.4.0}{
\subsection{encrqcs::qcs_run}{

\code{encrqcs::qcs_call} simplified. Instead of writing a .bit / .sh and
running that, it now directly calls Java via \code{system2}.
}

}

\subsection{News for version 0.3.0}{
\subsection{encrqcs}{

Bump dependencies.
}

}

\subsection{News for version 0.2.2}{
\subsection{encrqcs::qcs_run}{

\verb{[encrqcs::qcs_run]} works for the incidence dataset, but the Java
programme seems to fail for the other datasets at this time.
Potentially a protocol id problem.
}

}
}

