% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encrqcs.R
\docType{package}
\name{encrqcs}
\alias{encrqcs}
\title{\code{encrqcs}: JRC-ENCR QCS Tools}
\description{
encrqcs

Use JRC-ENCR QCS (https://encr.eu/tools-for-registries) from the
comfort of your R session.
\subsection{Installation}{

You can install the development version of encrqcs from \href{https://github.com/}{GitHub} with:\if{html}{\out{<div class="sourceCode r">}}\preformatted{# install.packages("devtools")
devtools::install_github("finnishcancerregistry/encrqcs")
}\if{html}{\out{</div>}}
}

\subsection{Example}{

Run incidence checks using JRC-ENCR QCS:\if{html}{\out{<div class="sourceCode r">}}\preformatted{results <- encrqcs::qcs(my_incidence, "incidence", "path/to/qcs/root/dir/")
}\if{html}{\out{</div>}}

Run mortality checks using JRC-ENCR QCS:\if{html}{\out{<div class="sourceCode r">}}\preformatted{results <- encrqcs::qcs(my_mortality, "mortality", "path/to/qcs/root/dir/")
}\if{html}{\out{</div>}}
}
}
\section{News}{

\subsection{News for version 0.2.2}{
\subsection{Object encrqcs::qcs_run news:}{

\verb{[encrqcs::qcs_run]} works for the incidence dataset, but the Java
programme seems to fail for the other datasets at this time.
Potentially a protocol id problem.
}

}
}

