% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{qcs_call}
\alias{qcs_call}
\title{Run JRC-ENCR QCS}
\usage{
qcs_call(
  dataset_file_path,
  qcs_dir_path,
  qcs_protocol_id,
  system2_arg_list = NULL,
  assertion_type = NULL
)
}
\arguments{
\item{dataset_file_path}{\verb{[character]} (no default)

Path to an existing file. This should be the dataset you want to use.}

\item{qcs_dir_path}{\verb{[character]} (no default)

Path to root directory of the JRC-ENCR QCS.}

\item{qcs_protocol_id}{\verb{[character, integer]} (no default)
\itemize{
\item \code{integer}: Used as-is. You have to study what the \code{.jar} does with that.
\item \code{character}: Take 11, 13, 15, or 17 for \code{"incidence"}, \code{"mortality"},
\code{"population"}, and \code{"lifetable"}, respectively.
}}

\item{system2_arg_list}{\verb{[NULL, list]} (default \code{NULL})

Optional, additional arguments passed to \verb{[system2]} if a list.}

\item{assertion_type}{See e.g. \verb{[dbc::report_to_assertion]}.}
}
\value{
The output of the \verb{[system2]} call is returned as-is.
}
\description{
Run JRC-ENCR QCS on a file on-disk.
}
