% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{qcs_call}
\alias{qcs_call}
\title{Run JRC-ENCR QCS}
\usage{
qcs_call(
  dataset_file_path,
  qcs_dir_path,
  qcs_protocol_id,
  system2_arg_list = NULL,
  assertion_type = NULL
)
}
\arguments{
\item{dataset_file_path}{\verb{[character]} (no default)

Path to an existing file. This should be the dataset you want to use.}

\item{qcs_dir_path}{\verb{[character]} (no default)

Path to root directory of the JRC-ENCR QCS.}

\item{qcs_protocol_id}{\verb{[character, integer]} (no default)
\itemize{
\item \code{integer}: Used as-is. You have to study what the \code{.jar} does with that.
\item \code{character}: Take 11, 13, 15, or 17 for \code{"incidence"}, \code{"mortality"},
\code{"population"}, and \code{"lifetable"}, respectively.
}}

\item{system2_arg_list}{\verb{[NULL, list]} (default \code{NULL})

Optional, additional arguments passed to \verb{[system2]} if a list.}

\item{assertion_type}{See e.g. \verb{[dbc::report_to_assertion]}.}
}
\value{
The output of the \verb{[system2]} call is returned as-is.
}
\description{
Run JRC-ENCR QCS on a file on-disk.
}
\details{
\itemize{
\item R working directory is temporarily set via \verb{[setwd]} to \code{qcs_dir_path}.
Your original working directory is always restored whether the call
succeeds or not.
\item The Java programme is called via a \code{system2} call. You may supply
arguments to it via \code{system2_arg_list} and you can even override
arguments determined by \code{encrqcs::qcs_call} --- any settings you pass
that way overwrite the defaults determine by \code{encrqcs::qcs_call}.
For instance, if you don't have the dir containing \code{java.exe} in your
system \code{PATH} environment variables, the default call will fail,
because it uses \code{command = "java"} --- but you can replace that with
a direct path to your \code{java.exe} via \code{system2_arg_list}.
\item If status code other than zero is reported, a warning is emitted by R.
E.g. status code 1 is "generic exit code" and means that the call
failed.
\item Finally,
The output of the \verb{[system2]} call is returned as-is.
}
}
