% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{qcs_read_results}
\alias{qcs_read_results}
\title{JRC-ENCR QCS Results}
\usage{
qcs_read_results(
  qcs_dir_path,
  qcs_protocol_id,
  fread_arg_list = NULL,
  readlines_arg_list = NULL,
  assertion_type = NULL,
  dataset_name = NULL
)
}
\arguments{
\item{qcs_dir_path}{\verb{[character]} (no default)

Path to root directory of the JRC-ENCR QCS.}

\item{qcs_protocol_id}{\verb{[character, integer]} (no default)
\itemize{
\item \code{integer}: Used as-is. You have to study what the \code{.jar} does with that.
\item \code{character}: Take 11, 13, 15, or 17 for \code{"incidence"}, \code{"mortality"},
\code{"population"}, and \code{"lifetable"}, respectively.
}}

\item{assertion_type}{See e.g. \verb{[dbc::report_to_assertion]}.}
}
\description{
Read JRC-ENCR QCS results into R.
}
