% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{qcs_read_results}
\alias{qcs_read_results}
\title{JRC-ENCR Results}
\usage{
qcs_read_results(
  qcs_dir_path,
  dataset_name,
  fread_arg_list = NULL,
  readlines_arg_list = NULL,
  assertion_type = "input"
)
}
\arguments{
\item{qcs_dir_path}{\verb{[character]} (no default)

Path to root directory of the JRC-ENCR QCS.}

\item{dataset_name}{\verb{[character]} (no default)

one of the following:
\itemize{
\item "incidence"
\item "mortality"
\item "population"
\item "lifetable"
}}

\item{fread_arg_list}{\verb{[NULL, list]} (default \code{NULL})

Additional arguments passed to \verb{[data.table::fread]} if a \code{list}.}

\item{readlines_arg_list}{\verb{[NULL, list]} (default \code{NULL})

Additional arguments passed to \verb{[readLines]} if a \code{list}.}

\item{assertion_type}{See e.g. \verb{[dbc::report_to_assertion]}.}
}
\description{
Read JRC-ENCR results into R.
}
