% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{qcs_read_results}
\alias{qcs_read_results}
\title{JRC-ENCR QCS Results}
\usage{
qcs_read_results(
  qcs_dir_path,
  dataset_name,
  fread_arg_list = NULL,
  readlines_arg_list = NULL,
  assertion_type = NULL
)
}
\arguments{
\item{qcs_dir_path}{\verb{[character]} (no default)

Path to root directory of the JRC-ENCR QCS.}

\item{dataset_name}{\verb{[character]} (no default)

one of the following:
\itemize{
\item "incidence"
\item "mortality"
\item "population"
\item "lifetable"
}}

\item{fread_arg_list}{\verb{[NULL, list]} (default \code{NULL})

Additional arguments passed to \verb{[data.table::fread]} if a \code{list}.
Argument \code{file} is determined internally and cannot be changed.}

\item{readlines_arg_list}{\verb{[NULL, list]} (default \code{NULL})

Additional arguments passed to \verb{[readLines]} if a \code{list}.
Argument \code{con} is determined internally and cannot be changed.}

\item{assertion_type}{See e.g. \verb{[dbc::report_to_assertion]}.}
}
\value{
The output will be a list with one
element for each result file. The name of the element is the name of
the file in the results dir.
}
\description{
Read JRC-ENCR QCS results into R.
}
\details{
\verb{[encrqcs::qcs_read_results]} performs the following steps:
\enumerate{
\item Directory containing results is assumed to look like e.g.
\verb{C:/path/to/qcs/output/incidence} for
\code{qcs_dir_path = "C:/path/to/qcs/"}, \code{dataset_name = "incidence"}.
\item Result files in the results dir are those that match case-insensitive
regex \code{"output([.]txt)|([.]csv)"}.
\item Each results file is read separately into R. Those ending in
\code{.txt} are read using \verb{[readLines]}. Those ending in \code{.csv} are read
using \verb{[data.table::fread]}.
The output will be a list with one
element for each result file. The name of the element is the name of
the file in the results dir.
}
}
