% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{qcs_run}
\alias{qcs_run}
\title{Run JRC-ENCR QCS}
\usage{
qcs_run(
  dataset_file_path,
  qcs_dir_path,
  qcs_protocol_id = 11L,
  system2_arg_list = NULL,
  assertion_type = "input"
)
}
\arguments{
\item{dataset_file_path}{\verb{[character]} (no default)

Path to an existing file. This should be the dataset you want to use.}

\item{qcs_dir_path}{\verb{[character]} (no default)

Path to root directory of the JRC-ENCR QCS.}

\item{qcs_protocol_id}{\verb{[integer]} (default \code{11L})

See the JRC-ENCR QCS User Compendium.}

\item{system2_arg_list}{\verb{[NULL, list]} (default \code{NULL})

Optional, additional arguments passed to \verb{[system2]} if a list.}

\item{assertion_type}{See e.g. \verb{[dbc::report_to_assertion]}.}
}
\description{
Run JRC-ENCR QCS on a file on-disk.
}
