% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{qcs_run}
\alias{qcs_run}
\title{JRC-ENCR QCS}
\usage{
qcs_run(
  dataset,
  dataset_name,
  qcs_dir_path,
  dataset_file_path = NULL,
  clean = NULL,
  write_arg_list = NULL,
  call_arg_list = NULL,
  read_arg_list = NULL,
  assertion_type = "input"
)
}
\arguments{
\item{dataset}{\verb{[data.frame]} (mandatory)

Column requirements vary depending on \code{dataset_name}.
Refer to the JRC-ENCR QCS User Compendium.}

\item{dataset_name}{\verb{[character]} (no default)

one of the following:
\itemize{
\item "incidence"
\item "mortality"
\item "population"
\item "lifetable"
}}

\item{qcs_dir_path}{\verb{[character]} (no default)

Path to root directory of the JRC-ENCR QCS.}

\item{dataset_file_path}{\verb{[NULL, character]} (default \code{NULL})

\code{dataset} needs to be written to hard drive for use by JRC-ENCR QCS.
You can (optionally) specify the path where to write \code{dataset} via this
arg.
\itemize{
\item \code{NULL}: \code{dataset} will be written into a temporary file given by
\code{tempfile(fileext = ".csv")}.
\item \code{character}: \code{dataset} will be written here.
}}

\item{clean}{\verb{[NULL, logical]} (default \code{NULL})

One of the following:
\itemize{
\item \code{NULL}     : Use \code{"both"}.
\item \code{"input"}  : Delete the file in \code{dataset_file_path} when this function
no longer needs it.
\item \code{"output"} : Delete output files when this function no longer need them.
Specifically, the whole output directory for the given
\code{dataset_name} is removed, e.g. \code{output/incidence},
in \code{qcs_dir_path}.
\item \code{"both"}   : Delete both input and output files/dirs.
\item \code{"neither"}: Don't delete anything.
}}

\item{write_arg_list}{\verb{[NULL, list]} (default \code{NULL})

Optional, additional arguments passed to \verb{[encrqcs::qcs_write_dataset]}
if a list. Arguments \code{dataset}, \code{dataset_name}, \code{file_path}, and
\code{assertion} type are determined internally and cannot be changed.}

\item{call_arg_list}{\verb{[NULL, list]} (default \code{NULL})

Optional, additional arguments passed to \verb{[encrqcs::qcs_call]}
if a list. Arguments \code{dataset_file_path}, \code{qcs_dir_path}, and
\code{assertion_type} are determined internally and
cannot be changed. \code{system2_arg_list} has a default which can be overridden.}

\item{read_arg_list}{\verb{[NULL, list]} (default \code{NULL})
Optional, additional arguments passed to \verb{[encrqcs::qcs_read_results]}
if a list. Arguments \code{qcs_dir_path}, \code{dataset_name}, and
\code{assertion_type} are determined internally and
cannot be changed.}

\item{assertion_type}{See e.g. \verb{[dbc::report_to_assertion]}.}
}
\value{
The output of \verb{[encrqcs::qcs_run]} is a list as returned by
\verb{[encrqcs::qcs_read_results]}
with the additional element \code{qcs_call_output}.
}
\description{
Run JRC-ENCR QCS and read results into R.
}
\details{
\verb{[encrqcs::qcs_run]} performs the following steps:
\enumerate{
\item \verb{[encrqcs::qcs_write_dataset]} is called to write \code{dataset} to disk
(see arg \code{write_arg_list}).
\item \verb{[encrqcs::qcs_call]} is called to run checks on the on-disk dataset
(see arg \code{call_arg_list}). Its output will be included in the output of
\code{encrqcs::qcs_run}.
\item \verb{[encrqcs::qcs_read_results]} is called to read results back into R.
\item The captured messages alluded to in step 2 are included in the output
as element named \code{qcs_call_output}.
The output of \verb{[encrqcs::qcs_run]} is a list as returned by
\verb{[encrqcs::qcs_read_results]}
with the additional element \code{qcs_call_output}.
\item Input / output files are removed on exit (whether successful or not)
of \verb{[encrqcs::qcs_run]} depending on arg \code{clean}.
}
}
\section{News for version 0.6.0}{


\code{encrqcs::qcs_run} now enforces the correct order of columns in \code{dataset}.

Replace output element \code{run_log} with \code{qcs_call_output}, the whole output
object of \code{encrqcs::qcs_call}.
}

\section{News for version 0.5.0}{


Removed override of \code{system2_arg_list} within \code{encrqcs::qcs_run}. So now
the user can pass \code{system2_arg_list} to the call of \code{encrqcs::qcs_call}.
}

\section{News for version 0.2.2}{


\verb{[encrqcs::qcs_run]} works for the incidence dataset, but the Java
programme seems to fail for the other datasets at this time.
Potentially a protocol id problem.
}

