% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{qcs_write_dataset}
\alias{qcs_write_dataset}
\title{JRC-ENCR QCS Input and Output}
\usage{
qcs_write_dataset(
  dataset,
  dataset_name,
  file_path,
  fwrite_arg_list = NULL,
  assertion_type = "input"
)
}
\arguments{
\item{dataset}{\verb{[data.frame]} (mandatory)

Column requirements vary depending on \code{dataset_name}.
Refer to the JRC-ENCR QCS User Compendium.}

\item{dataset_name}{\verb{[character]} (no default)

one of the following:
\itemize{
\item "incidence"
\item "mortality"
\item "population"
\item "lifetable"
}}

\item{file_path}{\verb{[character]} (no default)

path to write file to}

\item{fwrite_arg_list}{\verb{[NULL, list]} (default \code{NULL})

Additional arguments passed to \verb{[data.table::fwrite]} if a list. Arguments
\code{x}, \code{file}, and \code{sep} are determined internally and cannot be changed.}

\item{assertion_type}{See e.g. \verb{[dbc::report_to_assertion]}.}
}
\value{
\verb{[encrqcs::qcs_write_dataset]} always returns \code{NULL} invisibly. As a
side effect it writes \code{dataset} to disk.
}
\description{
Read and write .csv files for / created by the JRC-ENCR QCS.
}
\details{
\verb{[encrqcs::qcs_write_dataset]} performs simple checks on its input and
writes \code{dataset} on the disk using \verb{[data.table::fwrite]}.
}
