% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_qcs_file}
\alias{write_qcs_file}
\title{JRC-ENCR QCS Input and Output}
\usage{
write_qcs_file(
  dataset,
  dataset_name,
  file_path,
  fwrite_arg_list = NULL,
  assertion_type = "input"
)
}
\arguments{
\item{dataset}{\verb{[data.frame]} (mandatory)

Column requirements vary depending on \code{dataset_name}.
Refer to the JRC-ENCR QCS User Compendium.}

\item{file_path}{\verb{[character]} (no default)

path to write file to}

\item{fwrite_arg_list}{\verb{[NULL, list]} (default \code{NULL})

additional arguments passed to \link[data.table:fwrite]{data.table::fwrite}}
}
\description{
Read and write .csv files for / created by the JRC-ENCR QCS.
}
